requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai-whisper==20231117
httpx==0.27.0
python-multipart==0.0.9
torch==2.2.2
